Move Dashboard into presenters
