+++ spec/presenters/dashboard_spec.rb
@@ -2,7 +2,7 @@ require "spec_helper"
 require_relative "../../app/models/api_model"
 require_relative "../../app/models/solicitor"
 require_relative "../../app/models/organisation"
-require_relative "../../app/models/dashboard"
+require_relative "../../app/presenters/dashboard"
 
 RSpec.describe Dashboard, "#solicitors" do
   it "asks for all Solicitor objects" do
